moving ExperimentContainer into components folder
